additional conditions for mask creating (part2)

+++ pcrglobwb_utils/utils.py
@@ -88,7 +88,8 @@ def glob_folder(folder, grdc_column, verbose=False, encoding='ISO-8859-1'):
 
 def create_out_dir(out_dir):
 
-    shutil.rmtree(out_dir)
+    if os.path.isdir(out_dir):
+        shutil.rmtree(out_dir)
     
     os.makedirs(out_dir)
     click.echo('INFO -- saving output to folder {}'.format(out_dir))
